# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sous Mon Toit",
        "description": "Documentation for the API SousMonToit",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server of the real estate agency Sous Mon Toit"
        },
        {
            "url": "https://sousmontoit.herokuapp.com/",
            "description": "Server heroku of the real estate agency Sous Mon Toit"
        }
    ],
    "paths": {
        "/schedule/": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get list of all appointment",
                "description": "Return list of all appointments",
                "operationId": "getAppointmentList",
                "responses": {
                    "200": {
                        "description": "A list with appointment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointments"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment information",
                "description": "Return one appointment",
                "operationId": "getAppointment",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointments"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/customer/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments information for customer selected",
                "description": "Return all appointments for customer selected",
                "operationId": "getAppointmentCustomer",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointments"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/staff/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments information for staff selected",
                "description": "Return all appointments for staff selected",
                "operationId": "getAppointmentStaff",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointments"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/createAppt": {
            "post": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Store new appointment",
                "description": "Returns successful message",
                "operationId": "storeAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointments"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointments"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/update/{id}": {
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update existing appointment",
                "description": "Returns updated appointment data",
                "operationId": "updateAppointments",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointments"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointments"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedule/delete/{id}": {
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete existing appointment",
                "description": "Return succesfull message",
                "operationId": "deleteAppointment",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Appointments": {
                "title": "Appointments",
                "description": "Appointments model",
                "required": [
                    "id",
                    "scheduled_at",
                    "created_at",
                    "id_staff"
                ],
                "properties": {
                    "id": {
                        "description": "ID of appointment",
                        "format": "int64"
                    },
                    "scheduled_at": {
                        "description": "Date of appointment"
                    },
                    "created_at": {
                        "description": "Date of created"
                    },
                    "updated_at": {
                        "description": "Date of updated"
                    },
                    "notes": {
                        "description": "all you need to know for the appointment"
                    },
                    "id_estate": {
                        "description": "Id of the estate"
                    },
                    "id_staff": {
                        "description": "Id of the staff"
                    },
                    "id_customer": {
                        "description": "Id of the customer"
                    },
                    "id_appointment_type": {
                        "description": "Id of the appointment type"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "400": {
                "description": "Bad Request"
            },
            "401": {
                "description": "Access token is missing or invalid"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Resource Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource Not Found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "default": {
                "description": "une erreur \"inattendue\""
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Id of resource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token for customer or staff",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}